+++ CHANGELOG.md
@@ -3,7 +3,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
-## [Unreleased]
+## [Unreleased] = 2017-02-28
 ### Aded
 - check-zookeeper-reqs
 - check-zookeeper-latency